Remove SavedModel-specific methods from base Layer and Model, and add documentation for updating the serialization implementation.
PiperOrigin-RevId: 266978650
